requirements.txt
python-telegram-bot>=20.7
yt-dlp>=2023.12.30
python-dotenv>=1.0.0
requests>=2.31.0
instaloader>=4.10.1
